+++ ionoscloud/data_source_servers.go
@@ -47,7 +47,10 @@ func dataSourceServers() *schema.Resource {
 							Type:     schema.TypeString,
 							Optional: true,
 						},
-
+						"hostname": {
+							Type:     schema.TypeString,
+							Optional: true,
+						},
 						"cores": {
 							Type:     schema.TypeInt,
 							Computed: true,
@@ -368,6 +371,7 @@ func SetServerProperties(server ionoscloud.Server) map[string]interface{} {
 	if server.Properties != nil {
 		utils.SetPropWithNilCheck(serverMap, "template_uuid", server.Properties.TemplateUuid)
 		utils.SetPropWithNilCheck(serverMap, "name", server.Properties.Name)
+		utils.SetPropWithNilCheck(serverMap, "hostname", server.Properties.Hostname)
 		utils.SetPropWithNilCheck(serverMap, "cores", server.Properties.Cores)
 		utils.SetPropWithNilCheck(serverMap, "ram", server.Properties.Ram)
 		utils.SetPropWithNilCheck(serverMap, "availability_zone", server.Properties.AvailabilityZone)